Fix ParameterAutomation precision and duration underflow
- Promote startSampleTime and sampleRate from float to double so
  blockStartTime computation retains full precision against the
  engine's Float64 mSampleTime clock (float loses sub-sample
  accuracy after ~minutes of engine runtime)
- Fix unsigned underflow: duration adjustment for mid-block event
  entry now uses int64_t arithmetic before clamping; if the ramp
  is fully elapsed, the target value is applied immediately instead
  of scheduling a wrapped-around frame count

+++ Sources/SPFKAudioNodes/Parameters/NodeParameter+Automation.swift
@@ -90,8 +90,8 @@ extension NodeParameter {
             guard let observer = ParameterAutomationGetRenderObserver(
                 parameter.address,
                 avAudioNode.auAudioUnit.scheduleParameterBlock,
-                Float(sampleRate),
-                Float(startTime.sampleTime),
+                sampleRate,
+                Double(startTime.sampleTime),
                 automationBaseAddress,
                 events.count
             ) else {
